BackEnd/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
asyncio==4.0.0
base58==2.1.1
bcrypt==4.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coincurve==20.0.0
colorama==0.4.6
cryptography==46.0.6
cytoolz==1.1.0
dnspython==2.8.0
ecdsa==0.19.2
eth-hash==0.7.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
eventlet==0.40.4
fastapi==0.135.2
Flask==3.1.3
Flask-SocketIO==5.6.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
parsimonious==0.10.0
passlib==1.7.4
pefile==2024.8.26
psutil==7.2.2
psycopg2-binary==2.9.11
pyarmor==9.2.4
pyarmor.cli.core==8.1.0
pyasn1==0.6.3
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-engineio==4.13.1
python-jose==3.5.0
python-socketio==5.16.1
pywin32==311
pywin32-ctypes==0.2.3
regex==2026.2.28
requests==2.32.5
rsa==4.9.1
setuptools==82.0.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
toolz==1.1.0
tronpy==0.6.2
tronzap-sdk==1.0.7
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
wsproto==1.3.2
